+++ graphistry/plugins_types/graphviz_types.py
@@ -1,14 +1,16 @@
-
-from typing import Any, List, Set, Union, TYPE_CHECKING
+from typing import List, Set, Union, TYPE_CHECKING
 from typing_extensions import Literal
 
 if TYPE_CHECKING:
-    try:
-        from pygraphviz import AGraph
-    except:
-        AGraph: Any = None  # type: ignore
+    from IPython.display import SVG, Image
+    from pygraphviz import AGraph
+    PlotStaticResult = Union[bytes, str, SVG, Image]
 else:
-    AGraph: Any = None
+    class AGraph:
+        pass
+    PlotStaticResult = Union[bytes, str]
+
+GraphvizAttrValue = Union[str, int, float, bool]
 
 
 Prog = Literal[